pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "literary-finder"
version = "0.1.0"
description = "A Multi-Agent System for Deep Literary Discovery"
authors = [
    {name = "Pedro Acosta", email = "contact@pedroacosta.dev"},
]
readme = "README.md"
requires-python = ">=3.11"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "langchain>=0.1.0",
    "langgraph>=0.0.40",
    "langchain-openai>=0.1.0",
    "pydantic>=2.5.0",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "gradio>=3.50.0,<4.0.0",
    "httpx>=0.25.0",
    "openai>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.6.0",
    "pre-commit>=3.5.0",
]

[project.urls]
Homepage = "https://github.com/poacosta/literary-finder"
Documentation = "https://github.com/poacosta/literary-finder#readme"
Repository = "https://github.com/poacosta/literary-finder.git"
Issues = "https://github.com/poacosta/literary-finder/issues"

[project.scripts]
literary-finder = "literary_finder.app:main"
literary-finder-api = "literary_finder.api.server:main"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q"
testpaths = [
    "tests",
]
asyncio_mode = "auto"